+++ src/main/java/com/example/hotelproject/controller/OwnerController.java
@@ -1,12 +1,9 @@
 package com.example.hotelproject.controller;
 
-import com.example.hotelproject.controller.request.OwnerCreateRequest;
-import com.example.hotelproject.controller.request.OwnerUpdateRequest;
-import com.example.hotelproject.controller.response.HotelResponse;
-import com.example.hotelproject.controller.response.OwnerResponse;
-import com.example.hotelproject.controller.response.OwnersHotelsResponse;
+import com.example.hotelproject.controller.request.owner.OwnerCreateRequest;
+import com.example.hotelproject.controller.request.owner.OwnerUpdateRequest;
+import com.example.hotelproject.controller.response.owner.OwnerResponse;
 import com.example.hotelproject.service.OwnerService;
-import io.swagger.annotations.ApiOperation;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
