+++ src/main/java/com/example/hotelproject/controller/response/user/UserResponse.java
@@ -1,6 +1,5 @@
-package com.example.hotelproject.controller.response;
+package com.example.hotelproject.controller.response.user;
 
-import com.example.hotelproject.domain.Owner;
 import com.example.hotelproject.domain.User;
 import lombok.*;
