+++ src/main/java/com/example/hotelproject/domain/Hotel.java
@@ -1,6 +1,6 @@
 package com.example.hotelproject.domain;
 
-import com.example.hotelproject.controller.request.HotelUpdateRequest;
+import com.example.hotelproject.controller.request.hotel.HotelUpdateRequest;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -10,11 +10,9 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
 @Getter
