+++ src/main/java/com/example/hotelproject/repository/RoomRepository.java
@@ -1,7 +1,5 @@
 package com.example.hotelproject.repository;
 
-import com.example.hotelproject.controller.request.HotelCreateRequest;
-import com.example.hotelproject.domain.Hotel;
 import com.example.hotelproject.domain.room.Room;
 import java.util.List;
 import java.util.Optional;
