+++ src/main/java/com/example/hotelproject/service/UserService.java
@@ -1,9 +1,7 @@
 package com.example.hotelproject.service;
 
-import com.example.hotelproject.controller.request.UserCreateRequest;
-import com.example.hotelproject.controller.request.UserRegisterRequest;
-import com.example.hotelproject.controller.response.UserRegisterResponse;
-import com.example.hotelproject.controller.response.UserResponse;
+import com.example.hotelproject.controller.request.user.UserCreateRequest;
+import com.example.hotelproject.controller.response.user.UserResponse;
 import com.example.hotelproject.domain.User;
 import com.example.hotelproject.repository.UserRepository;
